projects/vision-agents/pyproject.toml
[project]
name = "agentic-vision"
version = "0.1.0"
description = "Agentic vision clients (Gemini code-execution Think-Act-Observe loops) and DB helpers"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "google-genai>=1.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "google-cloud-storage>=2.18.0",
    "pillow>=10.0.0",
    "dspy>=2.6.0",
    "loguru>=0.7.0",
    "numpy>=1.26.0",
    "opencv-python>=4.10.0",
    "openai>=1.0.0",
    "turbopuffer>=1.10.0",
    "modal>=0.73.0",
    "qdrant-client>=1.17.1",
    "scikit-image>=0.24.0",
    "transformers>=4.55.0",
    "torch>=2.8.0",
    "matplotlib>=3.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.6.0",
    "fiftyone>=1.14.1",
]

[project.scripts]
agentic-vision-viewer-api = "viewer_api.app:run"

[tool.setuptools.packages.find]
include = ["agentic_vision*", "viewer_api*"]

[dependency-groups]
dev = [
    "prek==0.3.8",
    "pytest==8.3.3",
    "ty==0.0.29",
]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["agentic_vision", "viewer_api", "scripts", "tests"]
exclude = [
    "agentic_vision/instance_segmentation/tools.py",
    "tests/test_instance_segmentation_refinement.py",
    "tests/test_standalone_instance_segmentation_tools.py",
]

[tool.ty.analysis]
allowed-unresolved-imports = [
    "sam3_inference.**",
    "api.db_gen.**",
    "api.services.**",
]
replace-imports-with-any = [
    "cv2",
    "cv2.**",
    "turbopuffer",
    "turbopuffer.**",
]

[tool.ty.rules]
unused-type-ignore-comment = "warn"
